Added logic for snapshot chunk 'basic' strategy
* Add Strategy (interface and implementation) for 'basic' snapshot chunks: chunks are generated by splitting a full encoded snapshot
* Add SnapshotPayload model to schema, since we are using it in two different packages
* Refactor snapshotmainblockservice to use new snapshot chunk strategy
* Refactor unit and integration tests

+++ core/service/snapshotBlockService.go
@@ -6,7 +6,7 @@ import (
 
 type (
 	SnapshotBlockServiceInterface interface {
-		NewSnapshotFile(block *model.Block, chunkSizeBytes int64) (*model.SnapshotFileInfo, error)
+		NewSnapshotFile(block *model.Block, chunkSizeBytes int) (*model.SnapshotFileInfo, error)
 		ImportSnapshotFile(snapshotFileInfo *model.SnapshotFileInfo) error
 		IsSnapshotHeight(height uint32) bool
 	}
